05-data-platforms/my-pipeline/pipeline/assets/ingestion/requirements.txt
# Dependencies for the `trips` ingestion Python asset
# Pin minimal, commonly used versions for reproducibility
pandas==2.2.3
requests==2.31.0
pyarrow==12.0.1
python-dateutil==2.8.2

